Design for Publications password page is done. Need to make the password protection logic. Starting with installation of vuex.

+++ database/migrations/2021_07_04_213359_create_people_members_table.php
@@ -15,7 +15,8 @@ class CreatePeopleMembersTable extends Migration
     {
         Schema::create('people_members', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('people_id');
+            $table->unsignedBigInteger('people_id');
+            // $table->foreign('people_id')->references('id')->on('people')->onDelete('cascade');
             $table->string('name');
             $table->string('designation');
             $table->string('cell_number');
